+++ public/index.html
@@ -1,0 +1,38 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <meta charset="UTF-8">
+  <title>AQLite</title>
+
+  <!-- Compiled and minified CSS -->
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
+  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.3/css/all.css"
+    integrity="sha384-UHRtZLI+pbxtHCWp1t77Bi1L4ZtiqrqD80Kn4Z8NTSRyMA2Fd33n5dQ8lWUE00s/" crossorigin="anonymous">
+  <link rel="stylesheet" href="css/styles.css">
+
+
+</head>
+
+<body>
+  <div class="flashContainer">
+    <object classid="clsid:D27CDB6E-AE6D-11cf-96B8-444553540000" id="externalInterface"
+      codebase="https://download.macromedia.com/pub/shockwave/cabs/flash/swflash.cab#version=9,0,28,0" height="100%"
+      width="100%">
+      <param name="movie" value="loader.swf" />
+      <param name="LOOP" value="false" />
+      <param name="SCALE" value="default" />
+      <param name="allowScriptAccess" value="always" />
+      <param name="allowFullScreen" value="true" />
+      <param name="menu" value="false" />
+      <param name="wmode" value="direct" />
+      <embed id="gameSWF" src="flash/aqlite/loader.swf" loop="false"
+        pluginspage="https://www.adobe.com/shockwave/download/download.cgi?P1_Prod_Version=ShockwaveFlash"
+        type="application/x-shockwave-flash" allowscriptaccess="always" allowFullScreen="true" scale="default"
+        wmode="direct" menu="false" height="100%" width="100%"></embed>
+    </object>
+  </div>
+</body>
+
+</html>
